% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myrror.utils.R
\name{apply_tolerance_colnames}
\alias{apply_tolerance_colnames}
\title{Apply Tolerance to Column Names}
\usage{
apply_tolerance_colnames(names, tolerance)
}
\arguments{
\item{names}{character vector, column names to be processed}

\item{tolerance}{character vector, options: 'no_cap', 'no_underscore', 'no_whitespace'}
}
\value{
a list of processed column names
}
\description{
Apply Tolerance to Column Names
}
\examples{
processed_names <- apply_tolerance_colnames(names(iris), tolerance = 'no_cap')
}
