% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myrror.utils.R
\name{check_df}
\alias{check_df}
\title{Check if the df arguments are valid, makes them into a data.frame if they are a list.}
\usage{
check_df(df)
}
\arguments{
\item{dfx}{data frame}
}
\description{
Check if the df arguments are valid, makes them into a data.frame if they are a list.
}
\examples{
check_dfs(iris, iris_var_v1)

}
