% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myrror.utils.R
\name{check_join_type}
\alias{check_join_type}
\title{Check join type. Internal function.}
\usage{
check_join_type(dfx, dfy, by.x, by.y, return_match = FALSE)
}
\arguments{
\item{dfx}{data.frame}

\item{dfy}{data.frame}

\item{by.x}{character vector, keys for by.y.}

\item{by.y}{character vector, keys for by.x.}

\item{return_match}{logical, default is FALSE.}
}
\value{
character/list depending on return_match FALSE/TRUE.
}
\description{
This function checks the join type between two data frames. Internal function.
It returns the type of match between the two data frames ("1:1", "1:m", "m:1", "m:m"),
and the identified and non-identified rows.
}
\keyword{internal}
