% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myrror.utils.R
\name{check_set_by}
\alias{check_set_by}
\title{Check if the df arguments are valid,
makes them into a data.frame if they are a list.}
\usage{
check_set_by(by = NULL, by.x = NULL, by.y = NULL)
}
\arguments{
\item{by}{character vector}

\item{by.x}{character vector}

\item{by.y}{character vector}
}
\description{
Check if the df arguments are valid,
makes them into a data.frame if they are a list.
}
\examples{

check_set_by(NULL, NULL, NULL) # rn set
check_set_by("id", NULL, NULL) # by set
check_set_by(NULL, "id", "id") # by.x and by.y set

}
