% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_type.R
\name{compare_type_int}
\alias{compare_type_int}
\title{Compare type of variables, internal function.}
\usage{
compare_type_int(myrror_object = NULL)
}
\arguments{
\item{myrror_object}{myrror object}
}
\value{
data.table object
}
\description{
Compare type of variables, internal function.
}
