% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_values.R
\name{compare_values}
\alias{compare_values}
\title{Compare values of matched datasets}
\usage{
compare_values(
  dfx = NULL,
  dfy = NULL,
  myrror_object = NULL,
  by = NULL,
  by.x = NULL,
  by.y = NULL,
  output = c("full", "simple", "silent"),
  interactive = getOption("myrror.interactive"),
  verbose = getOption("myrror.verbose"),
  tolerance = 1e-07
)
}
\arguments{
\item{dfx}{a non-empty data.frame.}

\item{dfy}{a non-empty data.frame.}

\item{myrror_object}{myrror object from \link{create_myrror_object}}

\item{by}{character, key to be used for dfx and dfy.}

\item{by.x}{character, key to be used for dfx.}

\item{by.y}{character, key to be used for dfy.}

\item{output}{character: one of "full", "simple", "silent"}

\item{interactive}{logical: If \code{TRUE}, print S3 method for myrror objects
displays by chunks. If \code{FALSE}, everything will be printed at once.}

\item{verbose}{logical: If \code{TRUE} additional information will be displayed}

\item{tolerance}{numeric, default to 1e-7.}
}
\value{
list object
}
\description{
Compare values of matched datasets
}
\examples{

comparison <- compare_values(iris, iris_var1)

}
