% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_myrror_object.R
\name{create_myrror_object}
\alias{create_myrror_object}
\title{Create Myrror Object}
\usage{
create_myrror_object()
}
\value{
list
}
\description{
Creates an empty list of class 'myrror'.
}
\examples{
new_myrror_object <- create_myrror_object()

new_myrror_object
}
