% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myrror.utils.R
\name{detect_sorting}
\alias{detect_sorting}
\title{Detect sorting in a data frame}
\usage{
detect_sorting(data)
}
\arguments{
\item{data}{data.frame}
}
\value{
list
}
\description{
Detect sorting in a data frame
}
\examples{
detect_sorting(iris)

}
