% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myrror.utils.R
\name{equal_with_tolerance}
\alias{equal_with_tolerance}
\title{Are these two values equal with tolerance applied?}
\usage{
equal_with_tolerance(x, y, tolerance = 1e-07)
}
\arguments{
\item{x}{numeric}

\item{y}{numeric}

\item{tolerance}{numeric}
}
\value{
logical
}
\description{
Are these two values equal with tolerance applied?
}
