% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_diff_values.R
\name{extract_diff_int}
\alias{extract_diff_int}
\title{Extract Different Values - Internal}
\usage{
extract_diff_int(myrror_object = NULL, tolerance = NULL)
}
\arguments{
\item{myrror_object}{myrror object}

\item{tolerance}{numeric, default to 1e-7}
}
\value{
list with two elements:
\enumerate{
\item diff_list
\item diff_table
}
}
\description{
Extract Different Values - Internal
}
