% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_diff_rows.R
\name{extract_diff_rows}
\alias{extract_diff_rows}
\title{Extract Different Rows - User-facing
Function to extract missing or new rows from comparing two dataframes.}
\usage{
extract_diff_rows(
  dfx = NULL,
  dfy = NULL,
  myrror_object = NULL,
  by = NULL,
  by.x = NULL,
  by.y = NULL,
  output = c("simple", "full", "silent"),
  tolerance = 1e-07,
  verbose = TRUE
)
}
\arguments{
\item{dfx}{a non-empty data.frame.}

\item{dfy}{a non-empty data.frame.}

\item{myrror_object}{myrror object from \link{create_myrror_object}}

\item{by}{character, key to be used for dfx and dfy.}

\item{by.x}{character, key to be used for dfx.}

\item{by.y}{character, key to be used for dfy.}

\item{output}{character: one of "full", "simple", "silent"}

\item{tolerance}{numeric, default to 1e-7.}

\item{verbose}{logical: If \code{TRUE} additional information will be displayed}
}
\value{
data.table object with the rows that are missing or new.
}
\description{
Extract Different Rows - User-facing
Function to extract missing or new rows from comparing two dataframes.
}
\examples{

extract_diff_rows(iris, iris_var1)
extract_diff_rows(survey_data, survey_data_2, by=c('country', 'year'))

}
