% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_diff_values.R
\name{extract_diff_values}
\alias{extract_diff_values}
\title{Extract Different Values - User-facing - List format
Function to extract rows with different values between two dataframes.}
\usage{
extract_diff_values(
  dfx = NULL,
  dfy = NULL,
  by = NULL,
  by.x = NULL,
  by.y = NULL,
  myrror_object = NULL,
  output = c("simple", "full", "silent"),
  tolerance = 1e-07,
  verbose = TRUE
)
}
\arguments{
\item{dfx}{data.frame object.}

\item{dfy}{data.frame object.}

\item{by}{character, key to be used for dfx and dfy.}

\item{by.x}{character, key to be used for dfx.}

\item{by.y}{character, key to be used for dfy.}

\item{myrror_object}{myrror object.}

\item{output}{character, one of "simple", "full", "silent".}

\item{tolerance}{numeric, default to 1e-7.}
}
\value{
list object with two items: diff_list and diff_table
}
\description{
Extract Different Values - User-facing - List format
Function to extract rows with different values between two dataframes.
}
\examples{

extract_diff_values(iris, iris_var1)
extract_diff_values(survey_data, survey_data_2, by=c('country', 'year'))

}
