% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myrror.utils.R
\name{get_correct_myrror_object}
\alias{get_correct_myrror_object}
\title{Get correct myrror object}
\usage{
get_correct_myrror_object(myrror_object, dfx, dfy, verbose)
}
\arguments{
\item{dfx}{a non-empty data.frame}

\item{dfy}{a non-empty data.frame}

\item{verbose}{logical: If \code{TRUE} additional information will be displayed}
}
\value{
myrror object
}
\description{
It checks all the arguments parsed to parent function. If
\code{myrror_object} if found, then it will be used. If not, it checks if both
databases are NULL. If they are it looks for the the last myrror object. If
nothing available, then error. Finally, it checks for the availability of
both datasets. If they are available, then create \code{myrror_object}
}
\keyword{internal}
