% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myrror.utils.R
\name{get_df_name}
\alias{get_df_name}
\title{Get the name of a data frame. Internal function.}
\usage{
get_df_name(df, original_call)
}
\arguments{
\item{df}{data.frame}

\item{original_call}{original call (df)}
}
\value{
character
}
\description{
This function gets the name of a data frame. Internal function.
If the data frame has a name attribute, it returns that. Otherwise, it returns
the deparse of the original call.
}
\keyword{internal}
