% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{iris_var1}
\alias{iris_var1}
\title{Iris Dataset Variation 1}
\format{
A data frame with 155 rows and 6 variables:
\describe{
\item{SL}{Numeric Sepal length in cm, with some NA values.}
\item{SW}{Numeric Sepal width in cm.}
\item{PL}{Numeric Petal length in cm.}
\item{PW}{Numeric Petal width in cm.}
\item{Species}{Factor w/ 3 levels "setosa","versicolor","virginica"}
\item{Petal.Area}{Numeric calculated as Petal.Length * Petal.Width.}
}
}
\source{
Modified iris dataset.
}
\usage{
iris_var1
}
\description{
This variation of the Iris dataset includes additional rows, a new column named \code{Petal.Area}
calculated as the product of \code{Petal.Length} and \code{Petal.Width}, and introduces \code{NA} values into the
\code{Sepal.Length} column. It tests the handling of extended datasets, new calculated fields, and missing values.
}
\keyword{datasets}
