% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{iris_var5}
\alias{iris_var5}
\title{Iris Dataset Variation 5}
\format{
A \code{data.frame} with 160 rows and 5 variables:
\describe{
\item{Sepal.Length}{Charactert, length in cm.}
\item{Sepal.Width}{Numeric, width in cm.}
\item{Petal.Length}{Numeric, length in cm.}
\item{Petal.Width}{Numeric, Petal width in cm.}
\item{Species}{Factor with levels: "setosa","versicolor","virginica".}
}
}
\source{
Modified \code{iris} dataset.
}
\usage{
iris_var5
}
\description{
This dataset variation includes:
\itemize{
\item Column with different type: \code{Sepal.Length} (character).
\item Column with different values: \code{Sepal.Length} (1 modified value).
}
}
\keyword{datasets}
