% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{iris_var6}
\alias{iris_var6}
\title{Iris Dataset Variation 6}
\format{
A \code{data.frame} with 146 rows and 5 variables:
\describe{
\item{Sepal.Length}{Numeric, Sepal length in cm.}
\item{Sepal.Width}{Numeric, Sepal width in cm.}
\item{Petal.Length}{Numeric, Petal length in cm.}
\item{Petal.Width}{Numeric, Petal width in cm.}
\item{Species}{Factor with levels: "setosa","versicolor","virginica".}
}
}
\source{
Modified \code{iris} dataset.
}
\usage{
iris_var6
}
\description{
Iris Dataset Variation 6
}
\keyword{datasets}
