% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myrror.utils.R
\name{is.sorted}
\alias{is.sorted}
\title{Check if a vector is sorted}
\usage{
is.sorted(x, ...)
}
\arguments{
\item{x}{vector}

\item{...}{additional arguments of is.sorted()}
}
\value{
logical
}
\description{
Check if a vector is sorted
}
\examples{
is.sorted(iris$Sepal.Length)

}
