% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myrror.utils.R
\name{is_dataframe_sorted_by}
\alias{is_dataframe_sorted_by}
\title{Detect sorting in a data frame}
\usage{
is_dataframe_sorted_by(df, by = NULL, decreasing = FALSE)
}
\arguments{
\item{df}{data.frame}

\item{by}{character vector}

\item{decreasing}{logical}
}
\value{
list
}
\description{
Detect sorting in a data frame
}
\examples{
is_dataframe_sorted_by(iris, by = "Sepal.Length")

}
