% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myrror.utils.R
\name{my_menu}
\alias{my_menu}
\title{Menu wrapper. Internal function.}
\usage{
my_menu(...)
}
\arguments{
\item{...}{arguments passed to the menu function.}
}
\value{
menu
}
\description{
This function is a wrapper around the base R menu function.
It is used to provide a consistent interface for the menu function.
}
\keyword{internal}
