% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myrror.utils.R
\name{my_readline}
\alias{my_readline}
\title{Readline wrapper. Internal function.}
\usage{
my_readline(...)
}
\arguments{
\item{...}{Arguments passed to \code{\link[base:readline]{base::readline()}}, particularly \code{prompt}.}
}
\value{
Character string containing the user's input.
}
\description{
This function is a wrapper around the base R readline function.
It is used to provide a consistent interface for the readline function.
}
\keyword{internal}
