% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myrror.utils.R
\name{pair_columns}
\alias{pair_columns}
\title{Pairs columns and prepares them for comparison.}
\usage{
pair_columns(merged_data_report, suffix_x = ".x", suffix_y = ".y")
}
\arguments{
\item{merged_data_report}{joined prepared_dfx and prepared_dfy.}

\item{suffix_x}{suffix for dfx (default .x)}

\item{suffix_y}{suffix for dfy (default .y)}
}
\value{
paired_columns
}
\description{
Pairs columns and prepares them for comparison.
}
\examples{
# Create a myrror object and extract column pairs
mo <- create_myrror_object(iris, iris_var1)
pair_columns(mo$merged_data_report)

}
