% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_dataset.R
\name{prepare_dataset}
\alias{prepare_dataset}
\title{Prepare Dataset
\itemize{
\item setDT
\item make valid col names
\item factor to character
\item keys in colnames
}}
\usage{
prepare_dataset(df, keys = NULL)
}
\arguments{
\item{df}{data.frame or data.table object}

\item{keys}{identifier}
}
\value{
a list
}
\description{
Prepare Dataset
\itemize{
\item setDT
\item make valid col names
\item factor to character
\item keys in colnames
}
}
\examples{
prepared_dataset <- prepare_dataset(iris_var1)
}
