% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myrror.utils.R
\name{prepare_df}
\alias{prepare_df}
\title{Prepares dataset for join, internal function.}
\usage{
prepare_df(df, by = NULL, factor_to_char = TRUE)
}
\arguments{
\item{df}{data.frame or data.table}

\item{by}{character vector}

\item{factor_to_char}{logical}
}
\description{
Prepares dataset for join, internal function.
}
\examples{
dataset <- data.frame(a = 1:10, b = letters[1:10])
prepare_df(dataset, by = "a")

}
