% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.myrror}
\alias{print.myrror}
\title{Print method for Myrror object.}
\usage{
\method{print}{myrror}(x, ...)
}
\arguments{
\item{x}{an object of class 'myrror_object'}

\item{...}{additional arguments}
}
\description{
Print method for Myrror object.
}
