% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suggested_ids.R
\name{suggested_ids}
\alias{suggested_ids}
\title{Suggested keys/ids}
\usage{
suggested_ids(df)
}
\arguments{
\item{df}{data.frame}
}
\value{
list with suggested keys/ids (first option and first combo)
}
\description{
Suggested keys/ids
}
