% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{survey_data}
\alias{survey_data}
\title{Survey Data
A country-year level dataset with 15 rows and 6 variables. 2 countries,
4 years, and 4 additional variables.}
\format{
A \code{data.table} with 16 rows and 4 variables:
\describe{
\item{country}{Factor with levels: "A", "B".}
\item{year}{Numeric, with values: 2010, 2011, 2012, 2013.}
\item{variable1}{Numeric.}
\item{variable2}{Numeric.}
\item{variable3}{Numeric.}
\item{variable4}{Numeric.}
}
@source Simulated data.
}
\usage{
survey_data
}
\description{
Survey Data
A country-year level dataset with 15 rows and 6 variables. 2 countries,
4 years, and 4 additional variables.
}
\keyword{datasets}
