% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{survey_data_2_cap}
\alias{survey_data_2_cap}
\title{Survey Data Variation 2 with Cap Keys}
\format{
A \code{data.table} with 15 rows and 6 variables:
\describe{
\item{country}{Factor with levels: "A", "B".}
\item{year}{Numeric, with values: 2010, 2011, 2012, 2013.}
\item{variable1}{Numeric.}
\item{variable2}{Numeric. Modified variable values.}
\item{variable3}{Numeric.}
\item{variable4}{Numeric.}
}
}
\source{
Simulated data.
}
\usage{
survey_data_2_cap
}
\description{
Survey Data Variation 2 with Cap Keys
}
\keyword{datasets}
