% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{survey_data_4}
\alias{survey_data_4}
\title{Survey Data Variation 4}
\format{
A \code{data.table} with 12 rows (4 missing) and 6 variables:
\describe{
\item{country}{Factor with levels: "A", "B".}
\item{year}{Numeric, with values: 2010-2017.}
\item{variable1}{Numeric.}
\item{variable2}{Numeric.}
\item{variable3}{Numeric.}
\item{variable4}{Numeric.}
}
}
\source{
Simulated data.
}
\usage{
survey_data_4
}
\description{
Survey Data Variation 4
}
\keyword{datasets}
