% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{survey_data_m1}
\alias{survey_data_m1}
\title{Survey Data m:1}
\format{
A \code{data.table} with 15 rows and 6 variables:
\describe{
\item{country}{Factor with levels: "A", "B".}
\item{year}{Numeric, with values: 2010-2017.}
\item{variable1}{Numeric.}
\item{variable2}{Numeric.}
\item{variable3}{Numeric.}
\item{variable4}{Numeric.}
}
}
\source{
Variation of \code{survey_data} with non-unique ids and a m:1 relationship between ids and values.
}
\usage{
survey_data_m1
}
\description{
Survey Data m:1
}
\keyword{datasets}
