% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_joyn.R
\name{temp_freq_table}
\alias{temp_freq_table}
\title{Tabulate simple frequencies}
\usage{
temp_freq_table(x, byvar, digits = 1, na.rm = FALSE, freq_var_name = "n")
}
\arguments{
\item{x}{data frame}

\item{byvar}{character: name of variable to tabulate. Use Standard evaluation.}

\item{digits}{numeric: number of decimal places to display. Default is 1.}

\item{na.rm}{logical: report NA values in frequencies. Default is FALSE.}

\item{freq_var_name}{character: name of the frequency variable. Default is "n".}
}
\value{
data.table with frequencies.
}
\description{
tabulate one variable frequencies
}
\keyword{internal}
