% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_joyn.R
\name{temp_is_id}
\alias{temp_is_id}
\title{Check if dt is uniquely identified by \code{by} variable}
\usage{
temp_is_id(
  dt,
  by,
  verbose = getOption("joyn.verbose", default = FALSE),
  return_report = FALSE
)
}
\arguments{
\item{dt}{either right of left table}

\item{by}{variable to merge by}

\item{verbose}{logical: if TRUE messages will be displayed}

\item{return_report}{logical: if TRUE, returns data with summary of duplicates.
If FALSE, returns logical value depending on whether \code{dt} is uniquely identified
by \code{by}}
}
\value{
logical or data.frame, depending on the value of argument \code{return_report}
}
\description{
report if dt is uniquely identified by \code{by} var or, if report = TRUE, the duplicates in \code{by} variable
}
\keyword{internal}
