% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_possible_ids.R
\name{temp_possible_ids}
\alias{temp_possible_ids}
\title{Temporary possible_ids() function}
\usage{
temp_possible_ids(df, max_combination_size = 5, verbose = FALSE)
}
\arguments{
\item{df}{data.frame}

\item{max_combination_size}{integer}

\item{verbose}{logical}
}
\value{
vector of possible ids
}
\description{
Temporary possible_ids() function
}
\keyword{internal}
