% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myrror.utils.R
\name{check_df}
\alias{check_df}
\title{Check if the df arguments are valid,
makes them into a data.frame if they are a list. Internal function.}
\usage{
check_df(df)
}
\arguments{
\item{df}{data frame}
}
\description{
Check if the df arguments are valid,
makes them into a data.frame if they are a list. Internal function.
}
\keyword{internal}
