% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_type.R
\name{compare_type}
\alias{compare_type}
\title{Compare type of variables}
\usage{
compare_type(
  dfx = NULL,
  dfy = NULL,
  by = NULL,
  by.x = NULL,
  by.y = NULL,
  myrror_object = NULL,
  output = c("full", "simple", "silent")
)
}
\arguments{
\item{dfx}{data.frame object}

\item{dfy}{data.frame object}

\item{myrror_object}{myrror object}

\item{output}{character, one of "full", "simple", "silent"}
}
\value{
list object
}
\description{
Compare type of variables
}
\examples{
comparison <- compare_type(iris, iris_var1)

}
