% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_type.R
\name{compare_type}
\alias{compare_type}
\title{Compare type of variables}
\usage{
compare_type(
  dfx = NULL,
  dfy = NULL,
  myrror_object = NULL,
  by = NULL,
  by.x = NULL,
  by.y = NULL,
  output = c("full", "simple", "silent"),
  interactive = getOption("myrror.interactive"),
  verbose = getOption("myrror.verbose")
)
}
\arguments{
\item{dfx}{a non-empty data.frame}

\item{dfy}{a non-empty data.frame}

\item{myrror_object}{myrror object}

\item{by}{character, key to be used for dfx and dfy}

\item{by.x}{character, key to be used for dfx}

\item{by.y}{character, key to be used for dfy}

\item{interactive}{logical: If \code{TRUE}, print S3 method for myrror objects
displays by chunks. If \code{FALSE}, everything will be printed at once}

\item{verbose}{logical: If \code{TRUE} additional information will be displayed}
}
\value{
list object
}
\description{
Compare type of variables
}
\examples{
comparison <- compare_type(iris, iris_var1)

}
