% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_values.R
\name{compare_values}
\alias{compare_values}
\title{Compare values of matched datasets}
\usage{
compare_values(
  dfx = NULL,
  dfy = NULL,
  by = NULL,
  by.x = NULL,
  by.y = NULL,
  myrror_object = NULL,
  output = c("full", "simple", "silent"),
  interactive = TRUE,
  tolerance = 1e-07
)
}
\arguments{
\item{dfx}{data.frame object.}

\item{dfy}{data.frame object.}

\item{by}{character, column name to join on.}

\item{by.x}{character, column name to join on in dfx.}

\item{by.y}{character, column name to join on in dfy.}

\item{myrror_object}{myrror object.}

\item{output}{character, one of "full", "simple", "silent".}

\item{interactive}{logical, default to TRUE.}

\item{tolerance}{numeric, default to 1e-7.}
}
\value{
list object
}
\description{
Compare values of matched datasets
}
\examples{

comparison <- compare_values(iris, iris_var1)

}
