% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_diff_rows.R
\name{extract_diff_rows}
\alias{extract_diff_rows}
\title{Extract Different Rows - User-facing
Function to extract missing or new rows from comparing two dataframes.}
\usage{
extract_diff_rows(
  dfx = NULL,
  dfy = NULL,
  by = NULL,
  by.x = NULL,
  by.y = NULL,
  myrror_object = NULL,
  output = c("simple", "full", "silent")
)
}
\arguments{
\item{dfx}{data.frame object.}

\item{dfy}{data.frame object.}

\item{by}{character, key to be used for dfx and dfy.}

\item{by.x}{character, key to be used for dfx.}

\item{by.y}{character, key to be used for dfy.}

\item{myrror_object}{myrror object.}

\item{output}{character, one of "simple", "full", "silent".}
}
\value{
data.table object with the rows that are missing or new.
}
\description{
Extract Different Rows - User-facing
Function to extract missing or new rows from comparing two dataframes.
}
\examples{

extract_diff_rows(iris, iris_var1)
extract_diff_rows(survey_data, survey_data_2, by=c('country', 'year'))
}
