% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_diff_values.R
\name{extract_diff_values}
\alias{extract_diff_values}
\title{Extract Different Values - User-facing
Function to extract rows with different values between two dataframes.}
\usage{
extract_diff_values(
  dfx = NULL,
  dfy = NULL,
  myrror_object = NULL,
  output = c("full", "simple", "silent")
)
}
\arguments{
\item{dfx}{data.frame object}

\item{dfy}{data.frame object}

\item{myrror_object}{myrror object}

\item{output}{character, one of "full", "simple", "silent"}
}
\value{
list object with two items: diff_list and diff_table
}
\description{
Extract Different Values - User-facing
Function to extract rows with different values between two dataframes.
}
\examples{

extract_diff_values(iris, iris_var1)

}
