% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myrror.utils.R
\name{get_correct_myrror_object}
\alias{get_correct_myrror_object}
\title{Get correct myrror object. Internal function.}
\usage{
get_correct_myrror_object(
  myrror_object,
  dfx,
  dfy,
  by,
  by.x,
  by.y,
  verbose,
  ...
)
}
\arguments{
\item{dfx}{a non-empty data.frame.}

\item{dfy}{a non-empty data.frame.}

\item{by}{character, key to be used for dfx and dfy.}

\item{by.x}{character, key to be used for dfx.}

\item{by.y}{character, key to be used for dfy.}

\item{verbose}{logical: If \code{TRUE} additional information will be displayed.}

\item{...}{other arguments parsed to parent function.}
}
\value{
myrror object
}
\description{
It checks all the arguments parsed to parent function. If
\code{myrror_object} if found, then it will be used. If not, it checks if both
databases are NULL. If they are it looks for the the last myrror object. If
nothing available, then error. Finally, it checks for the availability of
both datasets. If they are available, then create \code{myrror_object}
}
\keyword{internal}
