% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myrror.utils.R
\name{get_keys_or_default}
\alias{get_keys_or_default}
\title{Get keys or default}
\usage{
get_keys_or_default(keys, default = "rn")
}
\arguments{
\item{keys}{character vector}

\item{default}{character}
}
\value{
character
}
\description{
Get keys or default
}
