% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{iris_var1}
\alias{iris_var1}
\title{Iris Dataset Variation 1}
\format{
A \code{data.frame} with 155 rows and 6 variables:
\describe{
\item{Sepal.Length}{Numeric, Sepal length in cm, with some NA values.}
\item{Sepal.Width}{Numeric, Sepal width in cm.}
\item{Petal.Length}{Numeric, Petal length in cm.}
\item{Petal.Width}{Numeric, Petal width in cm.}
\item{Species}{Factor with levels: "setosa","versicolor","virginica".}
\item{Petal.Area}{Numeric, calculated as \code{Petal.Length * Petal.Width}.}
}
}
\source{
Modified \code{iris} dataset.
}
\usage{
iris_var1
}
\description{
This dataset variation includes:
\itemize{
\item Additional rows by duplicating the first 5 rows.
\item A new column \code{Petal.Area} calculated from \code{Petal.Length} and \code{Petal.Width}.
\item Introduction of NA values in \code{Sepal.Length}.
}
}
\details{
It tests the handling of extended datasets, new calculated fields, and missing values.
}
\keyword{datasets}
