% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{iris_var2}
\alias{iris_var2}
\title{Iris Dataset Variation 2}
\format{
A data frame with 150 rows and 5 variables, row order shuffled:
\describe{
\item{Sepal.Length}{Numeric, Sepal length in cm, modified by adding random values.}
\item{Sepal.Width}{Numeric, Sepal width in cm, with some NaN values.}
\item{Petal.Length}{Numeric, Petal length in cm.}
\item{Petal.Width}{Numeric, Petal width in cm.}
\item{Species}{Factor with levels: "setosa","versicolor","virginica".}
}
}
\source{
Modified \code{iris} dataset.
}
\usage{
iris_var2
}
\description{
This dataset variation includes:
\itemize{
\item NaN values in the \code{Sepal.Width} column.
\item Random adjustments to \code{Sepal.Length} to create a range of different values.
\item A shuffled order of rows to test comparison without reliance on row order.
}
}
\details{
It is designed to test the handling of \code{NaN} values, comparison of numeric differences, and insensitivity to row order.
}
\keyword{datasets}
