% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{iris_var3}
\alias{iris_var3}
\title{Iris Dataset Variation 3}
\format{
A data frame with 150 rows and 5 variables:
\describe{
\item{SL}{Character Sepal length in cm, originally numeric, with one NA value.}
\item{SW}{Mixed types. Numeric Sepal width > 3.5, otherwise "Wide".}
\item{PL}{Numeric Petal length in cm.}
\item{PW}{Numeric Petal width in cm.}
\item{Species}{Factor w/ 3 levels "setosa","versicolor","virginica"}
}
}
\source{
Modified iris dataset with column renaming and type conversion.
}
\usage{
iris_var3
}
\description{
This dataset variation tests column name changes, type conversion from numeric to character,
and the introduction of a missing value. The \code{Sepal.Length} column is renamed to \code{SL} and converted to
character type, with one NA value introduced. It assesses the package's ability to handle column renaming,
type mismatches, and missing data.
}
\keyword{datasets}
