% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myrror.utils.R
\name{my_menu}
\alias{my_menu}
\title{Menu wrapper. Internal function.}
\usage{
my_menu(...)
}
\arguments{
\item{...}{Arguments passed to \code{\link[utils:menu]{utils::menu()}}, including \code{choices} and \code{title}.}
}
\value{
Integer indicating the selected menu item.
}
\description{
This function is a wrapper around the base R menu function.
It is used to provide a consistent interface for the menu function.
}
\keyword{internal}
