% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myrror.utils.R
\name{my_readline}
\alias{my_readline}
\title{Readline wrapper. Internal function.}
\usage{
my_readline(...)
}
\arguments{
\item{...}{arguments passed to the readline function.}
}
\value{
readline
}
\description{
This function is a wrapper around the base R readline function.
It is used to provide a consistent interface for the readline function.
}
\keyword{internal}
