% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myrror-package.R
\docType{package}
\name{myrror-package}
\alias{myrror-package}
\title{myrror: Compare two Data Frames and Summarize Differences}
\description{
Convenience functions to compare two dataframes, explore, visualize and summarize similarities and differences.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://giorgiacek.github.io/myrror/}
}

}
\author{
\strong{Maintainer}: Giorgia Cecchinato \email{gcecchinato@worldbank.org}

}
\keyword{internal}
