% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myrror.R
\name{myrror}
\alias{myrror}
\alias{print.myrror_object}
\title{myrror main function}
\usage{
myrror(dfx, dfy, by = NULL, by.x = NULL, by.y = NULL, factor_to_char = TRUE)

\method{print}{myrror_object}(x, ...)
}
\arguments{
\item{dfx}{a non-empty data.frame}

\item{dfy}{a non-empty data.frame}

\item{by}{character, key to be used for dfx and dfy}

\item{by.x}{character, key to be used for dfx}

\item{by.y}{character, key to be used for dfy}

\item{factor_to_char}{TRUE or FALSE, default to TRUE.}

\item{x}{an object of class 'myrror_object'}

\item{...}{additional arguments}
}
\value{
draft: selection of by variables
}
\description{
myrror main function
}
\examples{
comparison <- myrror(iris, iris_var1)
}
