% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myrror.utils.R
\name{prepare_df}
\alias{prepare_df}
\title{Prepares dataset for joyn::joyn(). Internal function.}
\usage{
prepare_df(
  df,
  by = NULL,
  factor_to_char = TRUE,
  interactive = getOption("myrror.interactive")
)
}
\arguments{
\item{df}{data.frame or data.table}

\item{by}{character vector}

\item{factor_to_char}{logical}

\item{interactive}{logical}
}
\description{
Prepares dataset for joyn::joyn(). Internal function.
}
\examples{
# dataset <- data.frame(a = 1:10, b = letters[1:10])
# prepare_df(dataset, by = "a")

}
\keyword{internal}
