% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myrror.utils.R
\name{prepare_df}
\alias{prepare_df}
\title{Prepares dataset for joyn::joyn(). Internal function.}
\usage{
prepare_df(
  df,
  by = NULL,
  factor_to_char = TRUE,
  interactive = getOption("myrror.interactive"),
  verbose = getOption("myrror.verbose")
)
}
\arguments{
\item{df}{data.frame or data.table}

\item{by}{character vector}

\item{factor_to_char}{logical}

\item{interactive}{logical}

\item{verbose}{logical}
}
\value{
data.table
}
\description{
Prepares dataset for joyn::joyn(). Internal function.
}
\keyword{internal}
