% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myrror.utils.R
\name{prepare_df}
\alias{prepare_df}
\title{Prepares dataset for join}
\usage{
prepare_df(df, by, factor_to_char = TRUE)
}
\arguments{
\item{df}{data.frame or data.table}

\item{by}{character vector}

\item{factor_to_char}{logical}
}
\description{
Prepares dataset for join
}
\examples{
prepare_df(iris_var1, by = NULL) # adds "rn" variable

}
